Extract an NginxTest module from HttpBin
